# apps/hycanvas/backend/internal/httpapi/apidocs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HyCanvas Generation API",
    "version": "1.0.0",
    "description": "Programmatic deck generation and retrieval. Authenticate with a workspace API key (Authorization: Bearer hyk_...), minted by a workspace admin in Dashboard > Members > API keys. Keys are scoped (generate, read, export) and pinned to one workspace. Errors are RFC 7807 problem+json with a stable `code`. The same capabilities are exposed to AI agents over the Model Context Protocol at /mcp (streamable HTTP, same API keys/scopes/budgets)."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "hyk_..."
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "description": "Stable machine-readable error code."
          }
        }
      },
...
      "Job": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "completed",
              "failed"
            ]
          },
          "result": {
            "description": "On completion of a generation job: { designId, title, pageCount, editorUrl, images }.",
            "type": "object"
          },
          "error": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "apiKey": []
    }
  ],
  "paths": {
...
    "/jobs/{id}": {
      "get": {
        "summary": "Poll a job",
        "description": "Any valid key can poll jobs the API itself started (generation, document export). A completed generation job's result carries designId and editorUrl.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The job.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "404": {
            "description": "Unknown job (or another user's).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  }
}